refactor: remove call command, move to compat package
