+++ packages/__tests__/3-runtime-html/binding-commands.throttle-debounce.spec.ts
@@ -1,6 +1,6 @@
 import { TestContext, assert, createFixture } from '@aurelia/testing';
 import { BindingMode, customElement, bindable, Aurelia } from '@aurelia/runtime-html';
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 
 async function wait(ms: number): Promise<void> {
   await new Promise(resolve => setTimeout(resolve, ms));
@@ -355,7 +355,7 @@ describe('3-runtime-html/binding-commands.throttle-debounce.spec.ts', function (
 
         const component = new App();
 
-        ctx.container.register(delegateRegistration);
+        ctx.container.register(delegateSyntax);
 
         ctx.doc.body.appendChild(host);
         au.app({ component, host });
