+++ packages/__tests__/3-runtime-html/custom-elements.let.spec.ts
@@ -1,4 +1,4 @@
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 import { ILogger } from '@aurelia/kernel';
 import { assert, createFixture } from '@aurelia/testing';
 
@@ -14,7 +14,7 @@ describe('3-runtime-html/custom-elements.let.spec.ts', function () {
     'to-view',
   ]) {
     it(`throws on non .bind/.to-view: "${command}"`, async function () {
-      const { tearDown, start } = createFixture(`<let a.${command}="bc">`, class { }, [delegateRegistration], /* no start */false);
+      const { tearDown, start } = createFixture(`<let a.${command}="bc">`, class { }, [delegateSyntax], /* no start */false);
 
       let ex: Error;
       try {
