+++ packages/__tests__/3-runtime-html/dialog/dialog-service.spec.ts
@@ -1,4 +1,4 @@
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 import { noop } from '@aurelia/kernel';
 import {
   IDialogService,
@@ -874,7 +874,7 @@ describe('3-runtime-html/dialog/dialog-service.spec.ts', function () {
     for (const { title, only, afterStarted, afterTornDown } of testCases) {
       const $it = only ? it.only : it;
       $it(title, async function () {
-        const creationResult = createFixture('', class App { }, [delegateRegistration, DialogDefaultConfiguration]);
+        const creationResult = createFixture('', class App { }, [delegateSyntax, DialogDefaultConfiguration]);
         const { ctx, tearDown, startPromise } = creationResult;
         await startPromise;
         const dialogService = ctx.container.get(IDialogService);
