+++ packages/__tests__/3-runtime-html/enhance.spec.ts
@@ -17,7 +17,7 @@ import {
 } from '@aurelia/runtime-html';
 import { assert, TestContext, createFixture } from '@aurelia/testing';
 import { createSpecFunction, TestExecutionContext, TestFunction } from '../util.js';
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 
 describe('3-runtime/enhance.spec.ts', function () {
   interface TestSetupContext {
@@ -138,7 +138,7 @@ describe('3-runtime/enhance.spec.ts', function () {
       const container = ctx.container;
       const au = new Aurelia(container);
 
-      container.register(delegateRegistration);
+      container.register(delegateSyntax);
       let component;
       let dispose: () => void | Promise<void>;
       if (initialMethod === 'app') {
