+++ packages/__tests__/3-runtime-html/template-compiler.harmony.spec.ts
@@ -1,4 +1,4 @@
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 import {
   CustomAttribute,
   CustomElement,
@@ -253,7 +253,7 @@ describe('3-runtime-html/template-compiler.harmony.spec.ts', function () {
         <div click.bind="hasFocus2" click.capture="(log2 = (log2 || 0) + 1) && (hasFocus2 = log2)"></div>
       `,
       resources: [
-        delegateRegistration,
+        delegateSyntax,
         CustomAttribute.define(
           { name: 'click', bindables: ['value'] },
           class Click {
