+++ packages/__tests__/integration/app/startup.ts
@@ -5,7 +5,7 @@ import { appTemplate as template } from './app-template.js';
 import { atoms } from './atoms/index.js';
 import { callCollection } from './debug.js';
 import { MolecularConfiguration, molecules } from './molecules/index.js';
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { callSyntax, delegateSyntax } from '@aurelia/compat-v1';
 
 export class TestExecutionContext {
   public constructor(
@@ -31,7 +31,8 @@ export async function startup(config: StartupConfiguration = {}): Promise<TestEx
   au
     .register(
       atoms,
-      delegateRegistration,
+      delegateSyntax,
+      callSyntax,
       molecules.customize((molecularConfig: MolecularConfiguration) => {
         molecularConfig.useCSSModule = config.useCSSModule;
       }),
