+++ packages/__tests__/validation-html/validation-router.integration.spec.ts
@@ -1,4 +1,4 @@
-import { delegateRegistration } from '@aurelia/compat-v1';
+import { delegateSyntax } from '@aurelia/compat-v1';
 import {
   newInstanceForScope,
   Registration
@@ -120,7 +120,7 @@ describe('validation-html/validation-router.integration.spec.ts/integration', fu
         Registration.instance(ILocation, mockBrowserHistoryLocation),
         RouterConfiguration,
         ValidationHtmlConfiguration,
-        delegateRegistration,
+        delegateSyntax,
         ViewWithValidation,
         RedirectingView,
       )
